LSB Change Graph
Properly t the values for the x y and z coordinates of a selected site node.

Thresholds are still missing. (There is no found way to fill certain areas with different colors unless appending rectangles in the graph is to be done. Also when rectangles are appended, they do not accurately represent the thresholds for each graph)

Will see if adding threshold lines instead is much fit for use with these graphs.